moves button to components folder; adjusts loading state of button
