+++ src/components/Button.tsx
@@ -1,14 +1,21 @@
 import { FC, ReactNode } from 'react'
-import { TwStyle, styled } from 'twin.macro'
+import tw, { TwStyle, styled } from 'twin.macro'
 import 'styled-components/macro'
-import { Loader } from '../Loader'
+import { Loader } from './Loader'
 
 const BUTTON = styled.button<{ $cssStyle: TwStyle; $height: string; $width: string; $disabled: boolean }>`
+  ${tw`flex justify-center`}
   ${({ $cssStyle }) => $cssStyle};
   height: ${({ $height }) => $height};
   width: ${({ $width }) => $width};
   cursor: ${({ $disabled }) => ($disabled ? 'not-allowed' : 'pointer')};
 `
+
+const LOADER = () => (
+  <div tw="relative">
+    <Loader />
+  </div>
+)
 
 /* Pass Children, Width, Height as mandatory props to render the button 
 Pass all the css styles as tailwind classes in cssStyle prop */
@@ -20,10 +27,10 @@ export const Button: FC<{
   loading?: boolean
   disabled?: boolean
   [x: string]: any
-}> = ({ height, width, children, cssStyle, loading = false, disabled, ...props }) => (
+}> = ({ height, width, cssStyle, disabled, children, loading, ...props }) => (
   <>
     <BUTTON $cssStyle={cssStyle} $height={height} $width={width} $disabled={disabled} {...props}>
-      {loading ? <Loader /> : children}
+      {loading ? <LOADER /> : children}
     </BUTTON>
   </>
 )
